+++ paddle_billing/Entities/Metrics/MetricsInterval.py
@@ -1,0 +1,5 @@
+from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
+
+
+class MetricsInterval(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
+    Day: "MetricsInterval" = "day"